Fun Times With Scripts!
So, player can run around, sprint, and detect if there's an "object" around it. The detecting can either be done with a circle or polygon collider. The circle one is probably better for the pace of the game, since the arc where the player can't detect objects is so small as to be useless (plus then we get to spend less time mucking about with colliders). In future, the object will be pulled towards the player if it's within the collider and the player clicks the correct button (currently right mouse). Once it collides with the player, it will be added to the player's "inventory".

Yay
-The Sign Painter